+++ models/PostLinkcard.ts
@@ -1,0 +1,8 @@
+import type { PostImage } from '~/models/PostImage'
+
+export type PostLinkcard = {
+  url: string
+  thumbnail: PostImage
+  title: string
+  description: string
+}
